Hacky i18n implementation
Includes a copypasted version of MW's i18n implementation.

Parsing messages doesn't yet work, because it requires jqueryMsg. This
patch also adds module dependency handling (for dependencies that aren't
invoked through require(), which is needed to make jqueryMsg work). With
that and some more mocking of global objects, jqueryMsg's dependencies
all work, but jqueryMsg itself doesn't because it uses the DOM.

We could reimplement jqueryMsg in Node (possibly using domino), but it
also requires data from the server like SITENAME, in the form of a
virtual JSON file in the jqueryMsg module. Shipping that data by
executing the actual jqueryMsg code is appealing, but it might be hard
to make that code and its dependencies universal enough that they can
run in Node without a lot of fragile mocking.

What I'm planning to experiment with next is writing an implementation
of jqueryMsg that works both in the browser and in Node with domino.
Hopefully that way we won't have to run a bunch of MW code designed for
browsers in Node.